BEFORE SECURITY ADDITIONS
This is just the very basic bare minimum so far to the cloud system website. I already have a nearly impossible to breach strategy to ensure the passwords AND password hash patterns are ALL randomized and secured! I'm posting this commit now merely as a "backup plan" in case I lose the files or something haha.